fix(service): Change type data to Long and edit get Map localPath to url from clouldinary

+++ hrm/src/main/java/com/project/hrm/entities/Documents.java
@@ -18,7 +18,7 @@ public class Documents {
     private String description;
     private String filePath;
     private String fileType;
-    private Integer fileSize;
+    private Long fileSize;
     private LocalDateTime uploadDate;
 
     @Enumerated(EnumType.STRING)
